build(server index.js): modularized the main fetch function
