+++ server/index.js
@@ -6,9 +6,8 @@ const path = require('path');
 const fs = require('fs');
 const cors = require('cors');
 const axios = require('axios');
-const { GarminConnect } = require('garmin-connect');
-const garminCreds = require('../garmin.config.json');
-const date = require('date-and-time');
+
+const fetchNewActivities = require('./fetchNewActivities');
 
 const cron = require('node-cron');
 
@@ -50,35 +49,6 @@ app.listen(3001, () => {
 });
 
 
-
-
-async function fetchNewActivities() {
-
-  // Create a new Garmin Connect Client and login
-  let GCClient = new GarminConnect();
-  await GCClient.login(garminCreds.username, garminCreds.password);
-
-  // get last 5 activities
-  let activityList = await GCClient.getActivities(0, 5);
-
-  // get the activities named Wounded-Warrior
-  let activities = activityList.filter(activity => activity.activityName.includes('Wounded'));
-
-  // add wounded-warrior activities to the database
-  activities.forEach(activity => {
-    // `https://cryptographic.ninja:${port}/workouts-2021`
-    axios.post('http://localhost:3001/workouts-2021/', {
-      activityId: activity.activityId,
-      activityName: activity.activityName,
-      startTimeLocal: activity.startTimeLocal.substring(0, 10),
-      summarizedExerciseSets: activity.summarizedExerciseSets
-    })
-  });
-
-  const now = new Date();
-  const logString = date.format(now, 'MM/DD/YYYY hh:mm:ss A [MDT]');
-  console.log(`Fetched new activities @ ${logString}`);
-};
 
 fetchNewActivities();
